--- a/aten/src/ATen/native/Copy.cpp
+++ b/aten/src/ATen/native/Copy.cpp
@@ -166,6 +166,12 @@ static Tensor & copy_impl(Tensor & self, const Tensor & src, bool non_blocking) 
       }
       return self;
     }
+
+    if (fbgemm::fbgemmSupportedCPU() && fbgemm_copy_transpose_valid(self, src) &&
+      src.dtype() == self.dtype() && (src.dtype() == at::kFloat || src.dtype() == at::kBFloat16)) {
+      fbgemm_copy_transpose_same_type(self, src);
+      return self;
+    }
   #endif
 
   if (self.is_same(src)) {