@CaoE 2022-08-30
> @CaoE have you done any change to this diff compared with #83327?

Hi @frank-wei, I just add the check `src.size(src.dim() - 1) <= std::numeric_limits<int32_t>::max() &&
      src.size(src.dim() - 2) <= std::numeric_limits<int32_t>::max()` as fbgemm transpose has not made the data types of parameter consistent yet.  
```
template <typename T>
FBGEMM_API void transpose_simd(
    unsigned M,
    unsigned N,
    const T* src,
    unsigned ld_src,
    T* dst,
    unsigned ld_dst);

void transpose_avx512(
    int64_t M,
    int64_t N,
    const T* src,
    unsigned ld_src,
    T* dst,
    unsigned ld_dst);
```

Also still waiting for reproducing the internal errors.