@weiwangmeta 2022-09-20
Please also make sure the following reproducer passes of course since these two PRs are connected on this: 

https://github.com/pytorch/pytorch/pull/83327#issuecomment-1251703374