@CaoE 2022-09-23
> Please do not add `intel-priority` label on PRs which are yet not ready for review (i.e. in draft state or ones that can not pass CI)

Removed the label. This PR is not ready yet, because this PR has a large influence range, and the fbgemm's transpose is not optimal in some cases, requiring a lot of benchmark and model feedback results.